backend/toy_shop/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
coverage==6.5.0
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
Pillow==9.2.0
psycopg2==2.9.5
pytz==2022.5
requests==2.28.1
sqlparse==0.4.3
tzdata==2022.5
urllib3==1.26.12
